better helpers API
