+++ src/server/plugins/form-engine/fields/boolean-field.njk
@@ -1,13 +1,13 @@
 {% from "govuk/components/radios/macro.njk" import govukRadios %}
 
-{% macro booleanField(params) %}
+{% macro booleanField(params, helpers) %}
   {{ govukRadios({
     id: params.name,
     name: params.name,
     classes: "govuk-radios--inline govuk-radios--small",
     fieldset: {
       legend: {
-        text: params.resolveLabel(params.def, params.name),
+        text: helpers.resolveLabel(params.def, params.name),
         isPageHeading: false,
         classes: "govuk-fieldset__legend--m"
       }
