+++ src/server/plugins/form-engine/fields/radio-group-field.njk
@@ -1,13 +1,13 @@
 {% from "govuk/components/radios/macro.njk" import govukRadios %}
 
-{% macro radioGroupField(params) %}
+{% macro radioGroupField(params, helpers) %}
   {{ govukRadios({
     id: params.name,
     name: params.name,
     classes: "govuk-radios--small",
     fieldset: {
       legend: {
-        text: params.resolveLabel(params.def, params.name),
+        text: helpers.resolveLabel(params.def, params.name),
         isPageHeading: false,
         classes: "govuk-fieldset__legend--m"
       }
@@ -19,7 +19,7 @@
       html: params.def.flags.description,
       classes: "app-hint"
     },
-    items: params.resolveValid(params.def, params.formValues[params.name]),
+    items: helpers.resolveValid(params.def, params.formValues[params.name]),
     errorMessage: errorMessageHelper(params.formErrors[params.name].message)
   }) }}
 {% endmacro %}
