+++ src/server/plugins/form-engine/form.njk
@@ -21,13 +21,8 @@
           name: name,
           def: def,
           formValues: formValues,
-          formErrors: formErrors,
-          resolveLabel: resolveLabel,
-          resolveItems: resolveItems,
-          resolveMeta: resolveMeta,
-          resolveValid: resolveValid,
-          resolveSummaryRows: resolveSummaryRows
-        }) }}
+          formErrors: formErrors
+        }, helpers) }}
 
       {% endfor %}
